Require `pip>=9.0.0` in the dependent packages.
CentOS 7 still ships with `pip==1.4.1` which is too old
to correctly handle nested dependencies, i.e., dependencies of
packages that ElastiCluster depends upon.

Luckily `pip` seems to be able to update self on the fly,
although it's not clear whether it will re-execute self
with the new version while running, or if it must be re-run
in order for the changes to be picked up.